관리자 페이지 시작

+++ src/main/java/com/example/appserver/controller/TimerController.java
@@ -8,13 +8,13 @@ import org.springframework.web.bind.annotation.*;
 public class TimerController {
     private UserRepository UserRepository;
     @PutMapping("/timer/on/{id}") // 유저의 타이머 on
-    public void timerOn(@PathVariable("id") Long id,@RequestParam("status") boolean status) {
+    public void timerOn(@PathVariable("id") Integer id,@RequestParam("status") boolean status) {
         User user= UserRepository.findById(id);
         user.getTimer().timerOn();
 
     }
     @PutMapping("/timer/off/{id}") // 유저의 타이머 off
-    public void timerOff(@PathVariable("id") Long id,@RequestParam("status") boolean status) {
+    public void timerOff(@PathVariable("id") Integer id,@RequestParam("status") boolean status) {
         User user= UserRepository.findById(id);
         int gold=user.getTimer().timerOff();
         user.setGold(gold);
